New Custom Armour System
Includes a variety of changes largely to do with the new customizable armour system. Other minor tweaks to existing armour-related systems. Material cost standardization is a concern for immersive aspect of materials, game progression, and armour differentiation, but with a minor rebalance to my previous traits and material stats I think this is serviceable.